Separated build and release workflow for main, dev and other branches.
This commit modifies the GitHub Actions workflow to handle build and releases differently for main, dev and other branches.

The build-on-commit.yml workflow now triggers for all branches except main and dev ones, and only performs the build operation. In case of commits to main branch, release-on-main.yml is activated to build and release the software, tagging it as a regular release. New dev commits trigger build and pre-release via the new file prerelease-on-dev.yml. This change enables us to release pre-releases on dev branch allowing early adopters to test new features, while the stable version is only released on main branch.