Refactoring Github Actions workflows for efficient releases
Reconfigured the Github Actions workflows to separate the build and release processes for main, dev, and other branches to streamline our release process. We now utilize two separate workflows, 'release-on-main.yml' for the main branch and 'prerelease-on-dev.yml' for the dev branch. This change allows us to efficiently manage software releases, with pre-releases launched on the dev branch for early adopters to test new features while maintaining stability on the main branch.