Modify extraction of Maven project version in workflows
Changes have been made to how the Maven project version is extracted in both 'prerelease-on-dev.yml' and 'release-on-main.yml' workflows. The modification ensures the version is written into the Github environment to allow access in subsequent steps, primarily in the releasing stage. This way, the workflow is more efficient and doesn't need to extract the version multiple times.