Add AWS App Runner support and update workflows
This change introduces the App Runner service and adds it to the existing SQL-based Streaming Analytics data access configuration. It also separates the build action for Apache Flink Job and the data access UI in the GitHub action workflow for the dev branch pre-release. This separation helps with readability and maintainability of the code and CI/CD pipeline. It also updates the .gitignore file to ignore JS and typescript build files.

The CDK stack which can run on App Runner has been created with this commit as a faster, managed alternative to running the stack on a standalone machine. This can enhance the application performance by ensuring necessary resources are provisioned dynamically. It also added copyright notices to the data access configuration.