"Improve build system and deploy process for SQL based Streaming Analytics Data Access UI"
This commit involves wrapping the build and publish process of the SQL based Streaming Analytics Data Access UI in a GitHub workflow.

In the previous version, build artifacts were simply archived and available for download. Now, they're treated as pre-release assets on a GitHub repository which allows for better version and update management.

Furthermore, the change set also improves how the Docker container fetches the JAR file to be run. Rather than downloading from the release assets - a ZIP file needs to be handled - the JAR file directly fetched from the Artifacts.

In addition, a new Elastic Beanstalk stack is introduced to deploy the built JAR file. The AWS CDK was used to define this stack that would set up the needed resources in AWS.

Lastly, minor updates are also done on the package-lock.json file including addition of the 'request-promise' package.

This helps streamline the build and deployment process and is a step towards fully automated CI/CD.