Update GitHub workflow and application configurations
The GitHub workflow was updated to include an additional artifact named `sql-based-streaming-analytics-flink-job.jar` in the build-on-commit workflow. This change is necessary since we have added a new Flink job to the project.

The `application.properties` file was updated to remove a hardcoded value from the output stream name property. This property will now be populated dynamically at runtime.

The AWS CDK stack codes were also updated to make multiple modifications to meet the following requirements:
1. Update dependency names and versions in `package.json` and `package-lock.json`.
2. Replace the HTTPS library with the follow-redirects library for the file download method.
3. Refactor the application for Elastic Beanstalk and Kinesis Data Analytics.
4. Add a method to download the Flink job jar file from a specific GitHub release.
5. Add a SQL file for the Flink job.

These changes are part of an effort to improve codebase quality, functionality, and efficiency by using the latest and suitable libraries and organizing the codebase in a more modular way.