Add DynamoDB table for KCL checkpoints in stream analytics app
Introduced a DynamoDB table to function as a checkpoint for AWS KCL (Kinesis Client Library) within the SQL-based streaming analytics application. This way, we can have a persistent record of the data read by the KCL and prevent data loss in case of application failures.

Various corresponding changes were also made:
- Changes made to the `application.properties` file to add table name.
- Modified the CDK stack code to incorporate the creation of the DynamoDB table and granting necessary permissions.
- Finally, the changes were reflected in the DataAccessUiConfigurationProperties and KinesisConsumerService for accurate interfacing with the KCL and the DynamoDB table.

The addition of the table enhances data reliability and recovery in case of stream reading disruptions.