Update AWS solution stack version
The version of the solution stack on which the Elastic Beanstalk application runs has been updated in `full-solution-stack.ts`. This was changed from '64bit Amazon Linux 2023 v4.0.0 running Corretto 17' to '64bit Amazon Linux 2023 v4.0.1 running Corretto 17'. This upgrade ensures that the app will continue to receive support and patches beyond the end of life of the older version.