Modify jQuery CDN script tag for increased security
Added 'integrity' and 'crossorigin' attributes to the jQuery CDN link in the 'index.html' file of 'sql-based-streaming-analytics-data-access-ui'. This is done to enhance the script's Subresource Integrity (SRI) which adds integrity to the file loaded from the CDN, and prevents any tampering with the code. This increases the security of the application by ensuring that the browser only executes the scripts if the hashes match.