Add problem statement to README.md
Expanded the README.md with a detailed problem statement section. This provides context for users and helps them understand why using SQL-based streaming analytics with Apache Flink is beneficial, and what challenges it addresses. It also emphasizes how this solution broadens the field of streaming analytics to a wider audience including those proficient in SQL but not in the Apache Flink DataStreamAPI.