Update UI buttons according to app status and bump Spring Boot version
Modified the buttons in the index.html to reflect the actual application status. Now, the start button will only appear if the application status is 'ready', and the stop button will be visible otherwise. Also, we upgraded the version of Spring Boot Starter Parent in the pom.xml file to increase the security and performance of the application with recent bug fixes and updates in the new version.