Update request-promise to version 0.0.1 in package json
This change downgrades the version of the request-promise module from 4.2.6 to 0.0.1 in the package.json file according to the project requirements. As a result, some dependencies in the package-lock.json file were also updated. The reason for downgrading to a previous version is due to compatibility issues with the current system.