Remove repo and token parameters in workflows
The repo and token parameters were removed from the build-on-commit, prerelease-on-dev, and release-on-main GitHub workflows. This change was made as part of an effort to simplify our workflows and use defaults wherever possible. Furthermore, storing sensitive tokens in workflow files can potentially pose a security risk. Hence, its extraction was necessary.